+++ py/core/providers/embeddings/litellm.py
@@ -16,7 +16,8 @@ from core.base import (
     EmbeddingProvider,
     R2RException,
 )
-from core.utils import truncate_texts_to_token_limit
+
+from .utils import truncate_texts_to_token_limit
 
 logger = logging.getLogger()
