+++ py/core/utils/__init__.py
@@ -18,7 +18,6 @@ from shared.utils.base_utils import (
     increment_version,
     num_tokens,
     num_tokens_from_messages,
-    truncate_texts_to_token_limit,
     update_settings_from_dict,
     validate_uuid,
     yield_sse_event,
@@ -228,7 +227,6 @@ __all__ = [
     "convert_nonserializable_objects",
     "num_tokens",
     "num_tokens_from_messages",
-    "truncate_texts_to_token_limit",
     "SSEFormatter",
     "SearchResultsCollector",
     "update_settings_from_dict",
