+++ py/shared/utils/base_utils.py
@@ -9,7 +9,6 @@ from typing import TYPE_CHECKING, Any, Optional, Tuple, TypeVar
 from uuid import NAMESPACE_DNS, UUID, uuid4, uuid5
 
 import tiktoken
-from litellm import get_model_info, token_counter
 
 from ..abstractions import (
     AggregateSearchResult,
@@ -270,37 +269,6 @@ def num_tokens_from_messages(messages, model="gpt-4o"):
 
         tokens += 3  # every reply is primed with assistant
     return tokens
-
-
-def truncate_texts_to_token_limit(texts: list[str], model: str) -> list[str]:
-    """
-    Truncate texts to fit within the model's token limit.
-    """
-    try:
-        model_info = get_model_info(model=model)
-        if not model_info.get("max_input_tokens"):
-            return texts  # No truncation needed if no limit specified
-
-        truncated_texts = []
-        for text in texts:
-            text_tokens = token_counter(model=model, text=text)
-            assert model_info["max_input_tokens"]
-            if text_tokens > model_info["max_input_tokens"]:
-                estimated_chars = (
-                    model_info["max_input_tokens"] * 3
-                )  # Estimate 3 chars per token
-                truncated_text = text[:estimated_chars]
-                truncated_texts.append(truncated_text)
-                logger.warning(
-                    f"Truncated text from {text_tokens} to ~{model_info['max_input_tokens']} tokens"
-                )
-            else:
-                truncated_texts.append(text)
-
-        return truncated_texts
-    except Exception as e:
-        logger.warning(f"Failed to truncate texts: {str(e)}")
-        return texts  # Return original texts if truncation fails
 
 
 class SearchResultsCollector:
